# api/openapi-combined.yaml
openapi: 3.0.3
info:
  title: Git Simulation Platform API
  description: |
    Полное API платформы для моделирования Git с Code Review.
    
    ## Архитектура
    
    Платформа состоит из трёх микросервисов:
    
    ### 1. User Gateway Service (порт 8082)
    Фасад-сервис для конечных пользователей. Оркестрирует остальные сервисы,
    валидирует права доступа. **Рекомендуется использовать этот сервис.**
    
    Все запросы используют **имена** (team_name, username, repo_name, commit_name, pr_name).
    UUID/ID возвращаются только в ответах для справки.
    
    ### 2. PR Allocation Service (порт 8080)  
    Управление командами, пользователями и Pull Requests.
    Автоматическое назначение ревьюверов. (Внутренний API)
    
    ### 3. Code Storage Service (порт 8081)
    Хранение кода в виде деревьев коммитов. 
    Поддержка init, push, checkout, merge. (Внутренний API)
    
    ## Аутентификация
    
    Пользователь идентифицируется через заголовок `X-Username` (имя пользователя).

  version: "1.0.0"
  contact:
    name: API Support
servers:
  - url: http://localhost:8082
    description: User Gateway Service (рекомендуется)
  - url: http://localhost:8080
    description: PR Allocation Service (внутренний)
  - url: http://localhost:8081
    description: Code Storage Service (внутренний)
...
components:
  parameters:
...
    TeamIdQuery:
      name: team_id
      in: query
      required: true
      schema:
        type: string
        format: uuid
      description: UUID команды
    RootCommitQuery:
      name: root_commit
      in: query
      required: true
      schema:
        type: string
        format: uuid
      description: Корневой коммит репозитория
    CommitIdQuery:
      name: commit_id
      in: query
      required: true
      schema:
        type: string
        format: uuid
      description: Идентификатор коммита
  schemas:
...
    StorageCommit:
      type: object
      required: [commit_id, team_id, root_commit, parent_commit_ids, createdAt]
      properties:
        commit_id:
          type: string
          format: uuid
        team_id:
          type: string
          format: uuid
        root_commit:
          type: string
          format: uuid
        parent_commit_ids:
          type: array
          items:
            type: string
            format: uuid
        createdAt:
          type: string
          format: date-time
        commit_name:
          type: string
          nullable: true
...
paths:
...
  /storage/init:
    post:
      tags: ["[Storage] Commits"]
      summary: "[Internal] Инициализировать репозиторий"
      servers:
        - url: http://localhost:8081
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [team_id, code]
              properties:
                team_id:
                  type: string
                  format: uuid
                commit_name:
                  type: string
                code:
                  type: string
                  format: binary
      responses:
        '200':
          description: Репозиторий создан
          content:
            application/json:
              schema:
                type: object
                properties:
                  commit:
                    $ref: '#/components/schemas/StorageCommit'
  /storage/push:
    post:
      tags: ["[Storage] Commits"]
      summary: "[Internal] Создать коммит"
      servers:
        - url: http://localhost:8081
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [team_id, root_commit, commit_id, code]
              properties:
                team_id:
                  type: string
                  format: uuid
                root_commit:
                  type: string
                  format: uuid
                commit_id:
                  type: string
                  format: uuid
                  description: Parent commit ID
                commit_name:
                  type: string
                code:
                  type: string
                  format: binary
      responses:
        '201':
          description: Коммит создан
          content:
            application/json:
              schema:
                type: object
                properties:
                  commit:
                    $ref: '#/components/schemas/StorageCommit'
  /storage/checkout:
    get:
      tags: ["[Storage] Commits"]
      summary: "[Internal] Получить код коммита"
      servers:
        - url: http://localhost:8081
      parameters:
        - $ref: '#/components/parameters/TeamIdQuery'
        - $ref: '#/components/parameters/RootCommitQuery'
        - $ref: '#/components/parameters/CommitIdQuery'
      responses:
        '200':
          description: ZIP с кодом
          content:
            application/zip:
              schema:
                type: string
                format: binary
  /storage/merge:
    post:
      tags: ["[Storage] Commits"]
      summary: "[Internal] Смержить коммиты"
      servers:
        - url: http://localhost:8081
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [team_id, root_commit, commit_id1, commit_id2]
              properties:
                team_id:
                  type: string
                  format: uuid
                root_commit:
                  type: string
                  format: uuid
                commit_id1:
                  type: string
                  format: uuid
                commit_id2:
                  type: string
                  format: uuid
      responses:
        '201':
          description: Merge commit создан
          content:
            application/json:
              schema:
                type: object
                properties:
                  commit:
                    $ref: '#/components/schemas/StorageCommit'
...